Allow blocks to set $error and $redirect variables
These will trigger an error message to be shown to the user and
optionally a redirect to take place when these values are encountered
during an AJAX call into the block.

Example usage of the feature is to check whether appropriate $_SESSION
variables are set for use within a query. These might be absent during
an AJAX call if the client's session has expired. In that case an
appropriate message and a redirect to the login page can be sent to
the client.