Add feature to specify tables in YAML
BREAKING CHANGE
Tables can now also be specified in YAML format with a .yml file-extension.
These will be preferred over .php files if both are present. To further
improve readibility of the table specs, which is also the main reason to
offer YAML, we've stepped away from the '#1' array key pattern in favor
of more simple integer keys. This is unfortunately a breaking change, but
it seems worthwhile because it makes both the table specs and the code
easier to read. The latter might be faster too.

To update your table specs, just change all the array keys for options
such as 'sum' and 'edit' to their corresponding bare integers. So '#1'
=> true becomes 1 => true.