Convert float values back after PHP PDO has mangled them
Unfortunately PHP PDO has no transfer type for floats, so they will be turned
into strings. To enable proper sorting and filtering in libtables, the correct
datatype needs to be maintained. To facilitate this, I've made the lt_query()
function convert any floats back to the actual float type. This is not a very
neat solution but there is no alternative. This requires libtables to keep a
list of the native_type strings for floats of the various databases PDO
supports. So far only 'float4' has been added.