Add edit option onconflict
This is a bit PostgreSQL specific right now, since it triggers on the
string "duplicate key value" in an SQL error message. If such an error
occurs it executes the query (usually a SELECT returning an id) instead,
in a similar manner as UPSERT/ON CONFLICT works in plain SQL. General use-
case would be an insert across two tables, with the second containing the
id of the first as a foreign key. If the first insert fails, the onconflict-
query will retrieve the id of the existing record instead. This setup requires
the use of the 'keys' option in the edit block to ensure the proper sequence
for the inserts.